LayoutTests/editing/selection/ios/select-content-outside-and-inside-multicolumn-002.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
<meta charset="utf-8">

<head>
    <script src="../../../resources/ui-helper.js"></script>
    <script src="../../../resources/js-test.js"></script>
    <style>
        body {
            margin: 0;
        }

        #container {
            font-size: 16px;
            width: 310px;
        }
    </style>
    <script>
        jsTestIsAsync = true;

        addEventListener("load", async () => {
            description("When selecting content outside and inside multi-column, the highlight box should not extend to the unselected column.");

            var outside = document.getElementById("outside");
            var inside = document.getElementById("inside");
            await UIHelper.longPressElement(outside);
            await UIHelper.waitForSelectionToAppear();

            getSelection().setBaseAndExtent(outside, 0, inside, 1);
            await UIHelper.ensurePresentationUpdate();

            selectionRects = await UIHelper.getUISelectionViewRects();

            if (selectionRects.length != 4) {
                testFailed(`The value of selectionRects.length should be 4, but got ${selectionRects.length}`);
                finishJSTest();
                return;
            }
            testPassed(`selectionRects.length is ${selectionRects.length}`);

            shouldBe("selectionRects[0].width", "310");
            shouldBe("selectionRects[1].width", "310");
            // The highlight box does not extend to the second column.
            shouldBe("selectionRects[2].width", "150");
            shouldBeGreaterThan("150", "selectionRects[3].width");

            finishJSTest();
        });
    </script>
</head>

<body>
    <div id="container">
        <p id="outside">Here is the selected content outside of the multi-column element.</p>
        <div style="column-count: 2; column-gap: 10px;"><span id="inside">Here is the selected content at the beginning of multi-column.</span> Lorem ipsum dolor sit amet, consectetur adipiscing
            elit. Morbi mauris ex, finibus quis aliquam ac, interdum facilisis ligula.</div>
    </div>
</body>

</html>
